+++ test/App/Fossa/Configuration/ConfigurationSpec.hs
@@ -81,10 +81,12 @@ expectedReleaseGroupProject =
 expectedConfigProject :: ConfigProject
 expectedConfigProject =
   ConfigProject
-    { configProjID = Just "github.com/fossa-cli"
+    { configProjLocator = Just "custom+1/github.com/fossa-cli"
+    , configProjID = Just "github.com/fossa-cli"
     , configName = Just "fossa-cli"
     , configLink = Just "fossa.com"
     , configTeam = Just "fossa-team"
+    , configTeams = Just ["fossa-team"]
     , configJiraKey = Just "key"
     , configUrl = Just "fossa.com"
     , configPolicy = Just (PolicyName "license-policy")