+++ test/Test/MockApi.hs
@@ -240,10 +240,16 @@ matchExpectation a@(DeleteReleaseGroup{}) (ApiExpectation _ requestExpectation b
 matchExpectation a@(DeleteReleaseGroupRelease{}) (ApiExpectation _ requestExpectation b@(DeleteReleaseGroupRelease{}) resp) = checkResult requestExpectation a b resp
 matchExpectation a@(UpdateReleaseGroupRelease{}) (ApiExpectation _ requestExpectation b@(UpdateReleaseGroupRelease{}) resp) = checkResult requestExpectation a b resp
 matchExpectation a@(CreateReleaseGroup{}) (ApiExpectation _ requestExpectation b@(CreateReleaseGroup{}) resp) = checkResult requestExpectation a b resp
+matchExpectation a@(CreateReleaseGroupRelease{}) (ApiExpectation _ requestExpectation b@(CreateReleaseGroupRelease{}) resp) = checkResult requestExpectation a b resp
 matchExpectation a@(GetReleaseGroups{}) (ApiExpectation _ requestExpectation b@(GetReleaseGroups{}) resp) = checkResult requestExpectation a b resp
 matchExpectation a@(GetReleaseGroupReleases{}) (ApiExpectation _ requestExpectation b@(GetReleaseGroupReleases{}) resp) = checkResult requestExpectation a b resp
 matchExpectation a@(GetPolicies{}) (ApiExpectation _ requestExpectation b@(GetPolicies{}) resp) = checkResult requestExpectation a b resp
 matchExpectation a@(GetTeams{}) (ApiExpectation _ requestExpectation b@(GetTeams{}) resp) = checkResult requestExpectation a b resp
+matchExpectation a@(GetProjectV2{}) (ApiExpectation _ requestExpectation b@(GetProjectV2{}) resp) = checkResult requestExpectation a b resp
+matchExpectation a@(UpdateProject{}) (ApiExpectation _ requestExpectation b@(UpdateProject{}) resp) = checkResult requestExpectation a b resp
+matchExpectation a@(UpdateRevision{}) (ApiExpectation _ requestExpectation b@(UpdateRevision{}) resp) = checkResult requestExpectation a b resp
+matchExpectation a@(GetOrgLabels{}) (ApiExpectation _ requestExpectation b@(GetOrgLabels{}) resp) = checkResult requestExpectation a b resp
+matchExpectation a@(AddTeamProjects{}) (ApiExpectation _ requestExpectation b@(AddTeamProjects{}) resp) = checkResult requestExpectation a b resp
 matchExpectation _ _ = Nothing
 
 -- | Handles a request in the context of the mock API.